+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.7] - 2018-09-19
+### Added
+- Add babelPlugins as option in order to add babel plugins to default set
+
+### Changed
+- Re-added default istanbuljs babel plugins to plugin set
+
 ## [0.2.6] - 2018-09-10
 ### Added
 - Add support for `importMeta` instrumenting via Istanbul configuration